set description subcommand
